[MOD] Modified controls to prepare for platform change to mobile.
Changed the player controls and animation to function using a custom axis based on Unity's basic input system that allows the player to move left or right by holding their thumb on the corresponding side of the screen. I modified the level bouindaries to reflect the new default aspect ratio, and changed the text in the tutorial prompts to reflect the new control style. There are still several places the UI needs to be tweaked for this new aspect ratio, including titles and the return button in the menus.